apps/api/pyproject.toml
[project]
description = "Launch LMS Backend"
name = "launch-lms-api"
version = "1.0.0"
requires-python = ">=3.14.3,<3.14.4"
authors = [
    {name = "Badr B. (swve)"}
]
dependencies = [
    "boto3==1.42.74",
    "botocore==1.42.74",
    "faker==40.11.1",
    "fastapi==0.135.1",
    "httpx==0.28.1",
    "google-genai==1.68.0",
    "pwdlib[argon2]==0.3.0",
    "psycopg2-binary==2.9.11",
    "pydantic[email]>=2.12.5,<3.0.0",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "redis==7.3.0",
    "requests==2.32.5",
    "resend==2.26.0",
    "sqlmodel==0.0.37",
    "tiktoken==0.12.0",
    "uvicorn==0.42.0",
    "typer==0.24.1",
    "alembic==1.18.4",
    "alembic-postgresql-enum==1.10.0",
    "sqlalchemy-utils==0.42.1",
    "stripe==14.4.1",
    "pyjwt[crypto]==2.12.1",
    "beautifulsoup4==4.14.3",
    "pytest-asyncio==1.3.0",
    "sentry-sdk[fastapi]==2.55.0",
    "protobuf==7.34.1",
    "workos==5.46.0",
    "dnspython==2.8.0",
    "defusedxml==0.7.1",
    "llama-index-core==0.14.18",
    "pgvector==0.4.2",
    "pypdf==6.9.2",
]

[tool.ruff]
lint.ignore = ["E501", "E712"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark test as async",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    "ignore:.*crypt.*deprecated.*",
    "ignore:.*utcnow.*deprecated.*",
    "ignore:.*Instrumentation will have no effect.*",
    "ignore:.*handler names should be lower-case.*",
]
